backend/requirements-dev.txt
-r requirements.txt

ipython==5.1.0
django-extensions==1.7.3

#######################################################
### Debug                                           ###
#######################################################
django-debug-toolbar==1.5

django-dbshell-plus==1.0.1
ipdb==0.10.1

#######################################################
### Testing                                         ###
#######################################################
factory_boy==2.8.1
pytest==3.3.1
pytest-django==3.1.2

#######################################################
### Docs                                            ###
#######################################################
sphinx==1.4.8
# Allows 'Edit on github' button in Sphinx docs.
https://github.com/westurner/sphinxcontrib-srclinks/archive/master.zip
sphinx_rtd_theme==0.1.9
